Implementação da funcionalidade de calcular juros e calcular juros com formatação

+++ src/Api2/Diogo.Softplan.Challenge.Api2.Domain/Services/IApi1Service.cs
@@ -4,6 +4,6 @@ namespace Diogo.Softplan.Challenge.Api2.Domain.Services
 {
     public interface IApi1Service
     {
-        Task<double> ObterTaxaDeJurosAsync();
+        Task<decimal> ObterTaxaDeJurosAsync();
     }
 }
